Created an Accounts Section and merged CRUD functionality to handle all accounts. Added CRUD functionality to handle account roles as well.
