--- src/main/java/org/launchcode/buildMyAppTriangle_20/controllers/CustomerController.java
@@ -1,14 +1,0 @@
-package org.launchcode.buildMyAppTriangle_20.controllers;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("customers")
-public class CustomerController {
-    @GetMapping()
-    public String index() {
-        return "customers/index.html";
-    }
-}
