+++ src/main/resources/templates/accounts/delete.html
@@ -2,10 +2,10 @@
 <html lang="en" xmlns:th = "http://www.thymeleaf.org/">
 <head>
     <meta charset="UTF-8">
-    <title>Delete Employee</title>
+    <title>Delete Account</title>
 </head>
 <body>
-<p>Employee Delete Form!</p>
+<p>Account Delete Form!</p>
 <form method="post">
     <div>
         <div>
