+++ src/main/resources/templates/accounts/index.html
@@ -2,11 +2,11 @@
 <html lang="en" xmlns:th = "http://www.thymeleaf.org/">
 <head>
     <meta charset="UTF-8">
-    <title>Employees Main Index</title>
+    <title>Accounts Main Index</title>
 </head>
 <body>
-    <p>Employee Main Page!</p>
-    <p>Lists All Current Workers!</p>
+    <p>Account Main Page!</p>
+    <p>Lists All Current Accounts!</p>
     <table>
         <tr>
             <th>
