+++ src/main/resources/templates/accounts/view.html
@@ -2,7 +2,7 @@
 <html lang="en" xmlns:th = "http://www.thymeleaf.org/">
 <head>
     <meta charset="UTF-8">
-    <title>View Employee</title>
+    <title>View Account</title>
 </head>
 <body>
 <p th:text="${user.firstName}"></p>
