--- src/main/resources/templates/customers/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th = "http://www.thymeleaf.org/">
-<head>
-    <meta charset="UTF-8">
-    <title>Customers Main Index</title>
-</head>
-<body>
-    <p>Customer Main Page!</p>
-    <p>Details about websites, sign in and page links</p>
-</body>
-</html>
