--- src/main/resources/templates/employees/add.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Add Employee</title>
-</head>
-<body>
-    <p>Add Employee Page!</p>
-    <form method="post">
-        <div>
-            <label> Email : <input type="text" name="username"/></label>
-        </div>
-        <div>
-            <label> Password : <input type="password" name="password"/></label>
-        </div>
-        <div>
-            <label> First Name : <input type="text" name="firstName"/></label>
-        </div>
-        <div>
-            <label> Last Name : <input type="text" name="lastName"/></label>
-        </div>
-        <div>
-            <input type="submit" value="Add"/>
-        </div>
-    </form>
-</body>
-</html>
