--- src/main/resources/templates/employees/delete.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th = "http://www.thymeleaf.org/">
-<head>
-    <meta charset="UTF-8">
-    <title>Delete Employee</title>
-</head>
-<body>
-<p>Employee Delete Form!</p>
-<form method="post">
-    <div>
-        <div>
-            <label> Users :
-                <select name="employeeId">
-                    <optgroup label="Employees">
-                        <option th:each="employee: ${employees}" th:value="${employee.id}" th:text="${employee.username}"></option>
-                    </optgroup>
-                    <optgroup label="Customers">
-                        <option th:each="customer: ${customers}" th:value="${customer.id}" th:text="${customer.username}"></option>
-                    </optgroup>
-                </select>
-            </label>
-        </div>
-    </div>
-    <div>
-        <input type="submit" value="delete"/>
-    </div>
-</form>
-</body>
-</html>
