--- src/main/resources/templates/employees/index.html
@@ -1,77 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th = "http://www.thymeleaf.org/">
-<head>
-    <meta charset="UTF-8">
-    <title>Employees Main Index</title>
-</head>
-<body>
-    <p>Employee Main Page!</p>
-    <p>Lists All Current Workers!</p>
-    <table>
-        <tr>
-            <th>
-                Employee Id
-            </th>
-            <th>
-                First name
-            </th>
-            <th>
-                Last Name
-            </th>
-            <th>
-                Email
-            </th>
-        </tr>
-        <tr th:each="employee : ${employees}">
-            <td>
-                <a th:href="@{'/employees/view/' + ${employee.id}}" th:text="${employee.id}"></a>
-            </td>
-            <td>
-                <p th:text="${employee.firstName}"></p>
-            </td>
-            <td>
-                <p th:text="${employee.lastName}"></p>
-            </td>
-            <td>
-                <p th:text="${employee.username}"></p>
-            </td>
-            <td>
-
-            </td>
-        </tr>
-    </table>
-    <table>
-        <tr>
-            <th>
-                Customer Id
-            </th>
-            <th>
-                First name
-            </th>
-            <th>
-                Last Name
-            </th>
-            <th>
-                Email
-            </th>
-        </tr>
-        <tr th:each="customer : ${customers}">
-            <td>
-                <a th:href="@{'/customers/view/' + ${customer.id}}" th:text="${customer.id}"></a>
-            </td>
-            <td>
-                <p th:text="${customer.firstName}"></p>
-            </td>
-            <td>
-                <p th:text="${customer.lastName}"></p>
-            </td>
-            <td>
-                <p th:text="${customer.username}"></p>
-            </td>
-            <td>
-
-            </td>
-        </tr>
-    </table>
-</body>
-</html>
