--- src/main/resources/templates/employees/update.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th = "http://www.thymeleaf.org/">
-<head>
-    <meta charset="UTF-8">
-    <title>Update Employee</title>
-</head>
-<body>
-<form method="post">
-    <div>
-        <label> Email : <input type="text" name="username" th:value="${employee.username}"/></label>
-    </div>
-    <div>
-        <label> Password : <input type="password" name="password"/></label>
-    </div>
-    <div>
-        <label> First Name : <input type="text" name="firstName" th:value="${employee.firstName}"/></label>
-    </div>
-    <div>
-        <label> Last Name : <input type="text" name="lastName" th:value="${employee.lastName}"/></label>
-    </div>
-    <div>
-        <input type="submit" value="update"/>
-    </div>
-</form>
-</body>
-</html>
