--- src/main/resources/templates/employees/view.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th = "http://www.thymeleaf.org/">
-<head>
-    <meta charset="UTF-8">
-    <title>View Employee</title>
-</head>
-<body>
-<p th:text="${employee.firstName}"></p>
-<p th:text="${employee.lastName}"></p>
-<p th:text="${employee.username}"></p>
-</body>
-</html>
